Bunch 'o Deployment script improvements.

+++ Scripts/UninstallBaristaSearchService.ps1
@@ -4,20 +4,29 @@ write-host
 write-host "[[STEP]] Removing SharePoint Hooks" -foregroundcolor Yellow
 write-host 
 
+$ver = $host | select version
+if ($ver.Version.Major -gt 1)  {$Host.Runspace.ThreadOptions = "ReuseThread"}
+if ( (Get-PSSnapin -Name  Microsoft.SharePoint.PowerShell -ErrorAction SilentlyContinue) -eq $null )
+{
+    Add-PsSnapin Microsoft.SharePoint.PowerShell
+}
+
 $serviceLocation = Join-Path ([Microsoft.SharePoint.Utilities.SPUtility]::GetGenericSetupPath("ISAPI")) "BaristaServices\Search\SPBaristaSearchService.exe"
 
 # deactivate in SharePoint
-try {
-Remove-BaristaSearchService -ErrorAction SilentlyContinue
-}
-catch [System.Management.Automation.CommandNotFoundException] {
-  write-host 'Search Service Not Installed.' -foregroundcolor Yellow
+if (Get-Command "Remove-BaristaSearchService" -errorAction SilentlyContinue) {
+	try {
+		Remove-BaristaSearchService -ErrorAction SilentlyContinue
+	}
+	catch [System.Management.Automation.CommandNotFoundException] {
+	  write-host 'Search Service Not Installed.' -foregroundcolor Yellow
+	}
 }
 
 write-host 
 write-host "[[STEP]] Removing existing Search Service" -foregroundcolor Yellow
 write-host 
-$searchService = Get-WmiObject -Class Win32_Service -Filter "Name = 'BaristaSearchWindowsService'" -ComputerName $env:COMPUTERNAME
+$searchService = Get-WmiObject -Class Win32_Service -Filter "Name = 'BaristaSearchWindowsService'" -ComputerName $env:COMPUTERNAME | out-null
 if ($searchService -ne $null) 
 { 
 	$searchService.Delete()
